+++ ichnaea/scripts/tests/test_map.py
@@ -1,4 +1,3 @@
-from contextlib import contextmanager
 import os
 
 from mock import MagicMock, patch
@@ -6,25 +5,19 @@ from mock import MagicMock, patch
 from ichnaea.models.content import MapStat
 from ichnaea.scripts import map as scripts_map
 from ichnaea.scripts.map import (
-    export_to_csv,
-    generate,
+    export_file,
     main,
-    tempdir,
 )
-from ichnaea.tests.base import CeleryTestCase
+from ichnaea.tests.base import (
+    _make_db,
+    CeleryTestCase,
+)
 from ichnaea import util
-
-
-@contextmanager
-def mock_system_call():
-    mock_system = MagicMock()
-    with patch.object(scripts_map, 'system_call', mock_system):
-        yield mock_system
 
 
 class TestMap(CeleryTestCase):
 
-    def test_export_to_csv(self):
+    def test_export_file(self):
         today = util.utcnow().date()
         data = [
             MapStat(time=today, lat=12345, lon=12345),
@@ -37,8 +30,12 @@ class TestMap(CeleryTestCase):
 
         with util.selfdestruct_tempdir() as temp_dir:
             filename = os.path.join(temp_dir, 'map.csv.gz')
-            result = export_to_csv(self.session, filename)
+
+            result = export_file(
+                None, filename, -85051, 85052, -180000, 180001,
+                _db_rw=_make_db(), _session=self.session)
             self.assertEqual(result, 36)
+
             with util.gzip_open(filename, 'r') as fd:
                 written = fd.read()
             lines = [line.split(',') for line in written.split()]
@@ -48,45 +45,26 @@ class TestMap(CeleryTestCase):
             self.assertEqual(set([round(float(l[1]), 2) for l in lines]),
                              set([-11.0, 12.35]))
 
-    def test_generate(self):
-        with mock_system_call() as mock_system:
-            generate(self.db_rw, 's3_bucket',
-                     self.raven_client, self.stats_client,
-                     upload=False, concurrency=1, datamaps='')
-            mock_calls = mock_system.mock_calls
-            self.assertEqual(len(mock_calls), 3)
-            self.assertTrue('encode' in mock_calls[0][1][0])
-            self.assertTrue(mock_calls[1][1][0].startswith('enumerate'))
-            self.assertTrue(mock_calls[2][1][0].startswith('enumerate'))
-        self.check_stats(
-            timer=[('datamaps', ['count:csv_rows']),
-                   ('datamaps', ['func:encode']),
-                   ('datamaps', ['func:export_to_csv']),
-                   ('datamaps', ['func:render'])])
-
-    def test_generate_explicit_output(self):
-        with tempdir() as temp_dir:
-            with mock_system_call() as mock_system:
-                generate(self.db_rw, 's3_bucket',
-                         self.raven_client, self.stats_client,
-                         upload=False, concurrency=1,
-                         datamaps='', output=temp_dir)
-                mock_calls = mock_system.mock_calls
-                self.assertEqual(len(mock_calls), 3)
-
     def test_main(self):
-        with tempdir() as temp_dir:
-            with mock_system_call() as mock_system:
+        with util.selfdestruct_tempdir() as temp_dir:
+            mock_generate = MagicMock()
+            with patch.object(scripts_map, 'generate', mock_generate):
                 argv = [
                     'bin/location_map',
                     '--create',
+                    '--upload',
                     '--concurrency=1',
-                    '--datamaps=%s' % temp_dir,
+                    '--datamaps=%s/datamaps' % temp_dir,
                     '--output=%s' % temp_dir,
                 ]
                 main(argv,
-                     _db_rw=self.db_rw,
                      _raven_client=self.raven_client,
                      _stats_client=self.stats_client)
-                mock_calls = mock_system.mock_calls
-                self.assertEqual(len(mock_calls), 3)
+
+                self.assertEqual(len(mock_generate.mock_calls), 1)
+                args, kw = mock_generate.call_args
+
+                self.assertEqual(kw['concurrency'], 1)
+                self.assertEqual(kw['datamaps'], temp_dir + '/datamaps')
+                self.assertEqual(kw['output'], temp_dir)
+                self.assertEqual(kw['upload'], True)